+++ changelog.md
@@ -130,6 +130,7 @@ and prevent players from spawning a new space station at any claimed chunks.
 - Added Argonauts compat to space stations. Players on the same guild will be
 able to share space stations.
 - Sliding doors can now be redstone powered from any of its blocks, instead of just the bottom middle one.
+- Machines now work with comparators.
 
 ## Breaking Changes
 - Removed Glacian signs, and aeronos and strophar chests.
@@ -143,6 +144,3 @@ if you're using these recipes in modpacks, you'll need to change them. This incl
   - "ad_astra:oxygen_conversion" -> "ad_astra:oxygen_loading"
 - Removed the sky renderer JSON API. Sky renderers should now be done in code.
 - Removed the JSON APIs for adding planets, galaxies and info in the planets screen.
-
-## TODO:
-- Machines now work with comparators.